// lib/types.ts
export interface PortfolioRecord {
  id: string
  name: string
  age: number
  height: number
  weight: number
  haircolour: string
  eyecolour: string
  bust: string
  "breast-type": string
  Konfektion: number
  schoesize: number
  privatehair: string
  tattoo: string
  smoker: string
  nantionality: string
  escort: string
  langDE: string
  langFR: string
  langEN: string
  langIT: string
  langPL: string
  langRO: string
  langPT: string
  langNL: string
  langCN: string
  langOTHER: string
  descDE: string
  descEN: string
  descFR: string
  descIT: string
  serviceincl_kiss: string
  serviceincl_schmusen: string
  serviceincl_tonguekiss: string
  serviceincl_privatemassage: string
  serviceincl_b2bmassage: string
  serviceincl_erotic: string
  serviceincl_massage: string
  serviceincl_lickballs: string
  serviceincl_girlfriendsex: string
  serviceincl_sexinallpositions: string
  serviceincl_69: string
  serviceincl_Reizwäsche: string
  serviceincl_bathplay: string
  serviceincl_showerplay: string
  serviceincl_bjwith: string
  serviceincl_bjwithout: string
  serviceincl_deepthroat: string
  serviceincl_cumonbody: string
  serviceincl_cummultiple: string
  serviceincl_spanish: string
  serviceincl_strapon: string
  serviceincl_masturbation: string
  serviceincl_dildo: string
  serviceincl_games: string
  serviceincl_dirtytalk: string
  serviceincl_roleplay: string
  serviceincl_facesitting: string
  serviceincl_footfetish: string
  serviceincl_striptease: string
  serviceincl_passivtongueanal: string
  serviceextra_bjcuminmouth?: number
  serviceextra_bjcuminmouthswallow?: number
  serviceextra_cumonbody?: number
  serviceextra_natursektactive?: number
  photos?: Photo[]
}

export interface Photo {
  id: string
  url: string
  alt: string
  isPrimary: boolean
}